Refactor/rename package to lpg (#35)
* Refactor: Align package with FRD specification

This commit refactors the package to align with the name and structure specified in the Functional Requirements Document (FRD).

The key changes include:
- Renamed the package from `omop2neo4j` to `py-omop2neo4j-lpg`.
- Updated the source directory, `pyproject.toml`, and all relevant import statements and hardcoded strings.
- Repaired the test suite, which broke as a result of the rename, by fixing imports, mock patches, and logger configurations.

A thorough analysis of the existing codebase revealed that all other functional requirements from the FRD (including `concept_ancestor` handling and the `:Standard` label) were already fully implemented. This refactoring was the only change required to make the package compliant with the definitive specification.

* Fix: Correct package name to use underscore

Per user feedback, this commit changes the package name in `pyproject.toml` from `py-omop2neo4j-lpg` to `py_omop2neo4j_lpg`.

The script name and CLI help text have also been updated to match.
