+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "omop2neo4j"
-version = "0.1.0"
+name = "py_omop2neo4j_lpg"
+version = "0.2.0"
 authors = [
     { name="Gowtham Rao", email="rao@ohdsi.org" }
 ]
@@ -36,11 +36,14 @@ test = [
 ]
 
 [project.scripts]
-omop2neo4j = "omop2neo4j.cli:cli"
+py_omop2neo4j_lpg = "omop2neo4j_lpg.cli:cli"
 
 [project.urls]
-"Homepage" = "https://github.com/ohdsi/omop2neo4j" # Placeholder
-"Bug Tracker" = "https://github.com/ohdsi/omop2neo4j/issues" # Placeholder
+"Homepage" = "https://github.com/ohdsi/py_omop2neo4j_lpg"
+"Bug Tracker" = "https://github.com/ohdsi/py_omop2neo4j_lpg/issues"
+
+[tool.setuptools.packages.find]
+where = ["src"]
 
 [tool.pytest.ini_options]
 env_files = [
