--- src/omop2neo4j/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the omop2neo4j directory a Python package.
