+++ src/omop2neo4j_lpg/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the omop2neo4j_lpg directory a Python package.
