+++ tests/test_integration.py
@@ -2,8 +2,8 @@ import os
 import pytest
 from click.testing import CliRunner
 
-from omop2neo4j.cli import cli
-from omop2neo4j.config import settings
+from omop2neo4j_lpg.cli import cli
+from omop2neo4j_lpg.config import settings
 
 
 @pytest.fixture(scope="session")
